Implement Admin interface
- Created a session login based on an ENV variable
- moved administrative actions to admin namespace
- locked access from unauthenticated user
