+++ .gitignore
@@ -15,3 +15,7 @@
 /log/*
 !/log/.keep
 /tmp
+config/database.yml
+
+# Ignore application configuration
+/config/application.yml
