+++ Gemfile
@@ -13,6 +13,8 @@ gem 'autoprefixer-rails'
 gem 'redcarpet'
 gem 'puma'
 
+gem 'figaro'
+
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .coffee assets and views
