+++ app/views/layouts/application.html.erb
@@ -8,6 +8,7 @@
 </head>
 <body>
   <%= render 'shared/header' %>
+  <%= render 'shared/messages' %>
 
   <%= yield %>
