+++ app/views/shared/_messages.html.erb
@@ -1,6 +1,6 @@
 <% flash.each do |name, msg| %>
-  <div>(class="alert alert-#{name}")
-    <a class="close">(data-dismiss='alert') &#215;</a>
+  <div class="alert alert-<%= "#{name}" %>">
+    <a class="close" data-dismiss="alert">&#215;</a>
     <%= content_tag :div, msg, :id => "flash_#{name}" if msg.is_a?(String) %>
   </div>
 <% end %>
