+++ spec/support/macros.rb_spec.rb
@@ -1,0 +1,3 @@
+def sign_in
+  session[:admin] = true
+end
