+++ spec/support/shared_examples_spec.rb
@@ -1,0 +1,7 @@
+shared_examples "requires admin" do
+  it "redirects to root path" do
+    session[:admin] = nil
+    action
+    expect(response).to redirect_to root_path
+  end
+end
